% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ability.R
\name{ability}
\alias{ability}
\title{ability}
\usage{
ability(rank, nodePosition, score)
}
\arguments{
\item{rank}{This is the Rank of the maze.}

\item{nodePosition}{You need to calculate the nodePosition.}

\item{score}{This is the series of connected paths via the black dots.}
}
\description{
The ability function returns the weighted score of the individual given his raw score (i.e. the number of black dotes collected).
}
\details{
The ability function is calculated following the paper in XXXX.
}
\examples{
\dontrun{
 rank <-3
 score<- 8
 nodePosition <- colourNodePosition(rank=3,satPercent=0.5,seed=1)
 ability(rank,nodePosition,score)
}
}
\author{
Aiden Loe and Maria Sanchez
}

