% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genPathSeed.R
\name{genPathSeed}
\alias{genPathSeed}
\title{genUniqueSolution}
\usage{
genPathSeed(path, rank, satPercent, seed)
}
\arguments{
\item{path}{Selecting the specific number of paths to achieve the maximum score.}

\item{rank}{This is the rank of the maze.}

\item{satPercent}{This is of saturation percentage ranging from 0-1.}

\item{seed}{The starting seed to begin searching for the seed with specific paths.}
}
\description{
This generate the solution by searching for the SEED that returns the specific number of paths to achieve the maximum score for a given rank and saturation.
}
\details{
This might be computationally intensive as the maze size increases. The seed is necessary so that the algorithm does not always begin from the smallest value. Based on the seed value, it will search for the next biggest that returns 1 unique solution.
}
\examples{
rank <- 5
satPercent <- 0.5
seed <- 1

#Searches for just one unique solution
justOne <- genPathSeed(path=3,rank=rank,satPercent=satPercent,seed=seed)
nodePosition <- colourNodePosition(rank,satPercent,seed=justOne)
mazeEst(rank,nodePosition)
}
\author{
Aiden Loe and Maria Sanchez
}

