% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gridEighteenLeft}
\alias{gridEighteenLeft}
\title{Grid Eighteen Left}
\format{A data frame with 2 columns
\describe{
  \item{start}{Coordinates of Start Node.}
  \item{end}{Coordinates End Node.}
}}
\usage{
data(gridEighteenLeft)
}
\description{
This returns a eighteen grid left maze. These are standardized coordinates.
}
\examples{
\dontrun{

# Returns a Grid with rank = 18
data(gridEighteenLeft)
coordinates <- gridEighteenLeft

}

}
\keyword{datasets}
