% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gridEighteenRight}
\alias{gridEighteenRight}
\title{Grid Eighteen Right}
\format{A data frame with 2 columns
\describe{
  \item{start}{Coordinates of Start Node.}
  \item{end}{Coordinates End Node.}
}}
\usage{
data(gridEighteenRight)
}
\description{
This returns a eighteen grid right maze. These are standardized coordinates.
}
\examples{
\dontrun{

# Returns a Grid with rank = 18
data(gridEighteenRight)
coordinates <- gridEighteenRight

}

}
\keyword{datasets}
