% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gridEighteenUp}
\alias{gridEighteenUp}
\title{Grid Eighteen Up}
\format{A data frame with 2 columns
\describe{
  \item{start}{Coordinates of Start Node.}
  \item{end}{Coordinates End Node.}
}}
\usage{
data(gridEighteenUp)
}
\description{
This returns a eighteen grid upwards maze. These are standardized coordinates.
}
\examples{
\dontrun{

# Returns a Grid with rank = 18
data(gridEightteenUp)
coordinates <- gridEighteenUp

}

}
\keyword{datasets}
