% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gridEightLeft}
\alias{gridEightLeft}
\title{Grid Eight Left}
\format{A data frame with 2 columns
\describe{
  \item{start}{Coordinates of Start Node.}
  \item{end}{Coordinates End Node.}
}}
\usage{
data(gridEightLeft)
}
\description{
This returns a eight grid left maze. These are standardized coordinates.
}
\examples{
\dontrun{

# Returns a Grid with rank = 8
data(gridEightLeft)
coordinates <- gridEightLeft

}

}
\keyword{datasets}
