% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gridEightRight}
\alias{gridEightRight}
\title{Grid Eight Right}
\format{A data frame with 2 columns
\describe{
  \item{start}{Coordinates of Start Node.}
  \item{end}{Coordinates End Node.}
}}
\usage{
data(gridEightRight)
}
\description{
This returns a eight grid right maze. These are standardized coordinates.
}
\examples{
\dontrun{

# Returns a Grid with rank = 8
data(gridEightRight)
coordinates <- gridEightRight

}

}
\keyword{datasets}
