% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gridEightUp}
\alias{gridEightUp}
\title{Grid Eight Up}
\format{A data frame with 2 columns
\describe{
  \item{start}{Coordinates of Start Node.}
  \item{end}{Coordinates End Node.}
}}
\usage{
data(gridEightUp)
}
\description{
This returns a eight grid upwards maze. These are standardized coordinates.
}
\examples{
\dontrun{

# Returns a Grid with rank = 8
data(gridEightUp)
coordinates <- gridEightUp

}

}
\keyword{datasets}
