% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gridElevenDown}
\alias{gridElevenDown}
\title{Grid Eleven Down}
\format{A data frame with 2 columns
\describe{
  \item{start}{Coordinates of Start Node.}
  \item{end}{Coordinates End Node.}
}}
\usage{
data(gridElevenDown)
}
\description{
This returns a eleven grid downards maze. These are standardized coordinates.
}
\examples{
\dontrun{

# Returns a Grid with rank = 11
data(gridElevenDown)
coordinates <- gridElevenDown

}

}
\keyword{datasets}
