% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gridElevenLeft}
\alias{gridElevenLeft}
\title{Grid Eleven Left}
\format{A data frame with 2 columns
\describe{
  \item{start}{Coordinates of Start Node.}
  \item{end}{Coordinates End Node.}
}}
\usage{
data(gridElevenLeft)
}
\description{
This returns a eleven grid left maze. These are standardized coordinates.
}
\examples{
\dontrun{

# Returns a Grid with rank = 11
data(gridElevenLeft)
coordinates <- gridElevenLeft

}

}
\keyword{datasets}
