% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gridElevenRight}
\alias{gridElevenRight}
\title{Grid Eleven Right}
\format{A data frame with 2 columns
\describe{
  \item{start}{Coordinates of Start Node.}
  \item{end}{Coordinates End Node.}
}}
\usage{
data(gridElevenRight)
}
\description{
This returns a eleven grid right maze. These are standardized coordinates.
}
\examples{
\dontrun{

# Returns a Grid with rank = 11
data(gridElevenRight)
coordinates <- gridElevenRight

}

}
\keyword{datasets}
