% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gridElevenUp}
\alias{gridElevenUp}
\title{Grid Eleven Up}
\format{A data frame with 2 columns
\describe{
  \item{start}{Coordinates of Start Node.}
  \item{end}{Coordinates End Node.}
}}
\usage{
data(gridElevenUp)
}
\description{
This returns a eleven grid upwards maze. These are standardized coordinates.
}
\examples{
\dontrun{

# Returns a Grid with rank = 11
data(gridElevenUp)
coordinates <- gridElevenUp

}

}
\keyword{datasets}
