% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gridFifteenDown}
\alias{gridFifteenDown}
\title{Grid Fifteen Down}
\format{A data frame with 2 columns
\describe{
  \item{start}{Coordinates of Start Node.}
  \item{end}{Coordinates End Node.}
}}
\usage{
data(gridFifteenDown)
}
\description{
This returns a fifteen grid downards maze. These are standardized coordinates.
}
\examples{
\dontrun{

# Returns a Grid with rank = 15
data(gridFifteenDown)
coordinates <- gridFifteenDown

}

}
\keyword{datasets}
