% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gridFifteenLeft}
\alias{gridFifteenLeft}
\title{Grid Fifteen Left}
\format{A data frame with 2 columns
\describe{
  \item{start}{Coordinates of Start Node.}
  \item{end}{Coordinates End Node.}
}}
\usage{
data(gridFifteenLeft)
}
\description{
This returns a fifteen grid left maze. These are standardized coordinates.
}
\examples{
\dontrun{

# Returns a Grid with rank = 15
data(gridFifteenLeft)
coordinates <- gridFifteenLeft

}

}
\keyword{datasets}
