% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gridFifteenRight}
\alias{gridFifteenRight}
\title{Grid Fifteen Right}
\format{A data frame with 2 columns
\describe{
  \item{start}{Coordinates of Start Node.}
  \item{end}{Coordinates End Node.}
}}
\usage{
data(gridFifteenRight)
}
\description{
This returns a fifteen grid right maze. These are standardized coordinates.
}
\examples{
\dontrun{

# Returns a Grid with rank = 15
data(gridFifteenRight)
coordinates <- gridFifteenRight

}

}
\keyword{datasets}
