% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gridFiveDown}
\alias{gridFiveDown}
\title{Grid Five Down}
\format{A data frame with 2 columns
\describe{
  \item{start}{Coordinates of Start Node.}
  \item{end}{Coordinates End Node.}
}}
\usage{
data(gridFiveDown)
}
\description{
This returns a five grid downwards maze. These are standardized coordinates.
}
\examples{
\dontrun{

# Returns a Grid with rank = 5
data(gridFiveDown)
coordinates <- gridFiveDown

}

}
\keyword{datasets}
