% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gridFiveLeft}
\alias{gridFiveLeft}
\title{Grid Five Left}
\format{A data frame with 2 columns
\describe{
  \item{start}{Coordinates of Start Node.}
  \item{end}{Coordinates End Node.}
}}
\usage{
data(gridFiveLeft)
}
\description{
This returns a five grid left maze. These are standardized coordinates.
}
\examples{
\dontrun{

# Returns a Grid with rank = 5
data(gridFiveLeft)
coordinates <- gridFiveLeft

}

}
\keyword{datasets}
