% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gridFiveRight}
\alias{gridFiveRight}
\title{Grid Five Right}
\format{A data frame with 2 columns
\describe{
  \item{start}{Coordinates of Start Node.}
  \item{end}{Coordinates End Node.}
}}
\usage{
data(gridFiveRight)
}
\description{
This returns a five grid right maze. These are standardized coordinates.
}
\examples{
\dontrun{

# Returns a Grid with rank = 5
data(gridFiveRight)
coordinates <- gridFiveRight

}

}
\keyword{datasets}
