% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gridFiveUp}
\alias{gridFiveUp}
\title{Grid Five Up}
\format{A data frame with 2 columns
\describe{
  \item{start}{Coordinates of Start Node.}
  \item{end}{Coordinates End Node.}
}}
\usage{
data(gridFiveUp)
}
\description{
This returns a five grid upwards maze. These are standardized coordinates.
}
\examples{
\dontrun{

# Returns a Grid with rank = 5
data(gridFiveUp)
coordinates <- gridFiveUp

}

}
\keyword{datasets}
