% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gridFourDown}
\alias{gridFourDown}
\title{Grid Four Down}
\format{A data frame with 2 columns
\describe{
  \item{start}{Coordinates of Start Node.}
  \item{end}{Coordinates End Node.}
}}
\usage{
data(gridFourDown)
}
\description{
This returns a four grid downwards maze. These are standardized coordinates.
}
\examples{
\dontrun{

# Returns a Grid with rank = 4
data(gridFourDown)
coordinates <- gridFourDown

}

}
\keyword{datasets}
