% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gridFourLeft}
\alias{gridFourLeft}
\title{Grid Four Left}
\format{A data frame with 2 columns
\describe{
  \item{start}{Coordinates of Start Node.}
  \item{end}{Coordinates End Node.}
}}
\usage{
data(gridFourLeft)
}
\description{
This returns a four grid left maze. These are standardized coordinates.
}
\examples{
\dontrun{

# Returns a Grid with rank = 4
data(gridFourLeft)
coordinates <- gridFourLeft

}

}
\keyword{datasets}
