% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gridFourRight}
\alias{gridFourRight}
\title{Grid Four Right}
\format{A data frame with 2 columns
\describe{
  \item{start}{Coordinates of Start Node.}
  \item{end}{Coordinates End Node.}
}}
\usage{
data(gridFourRight)
}
\description{
This returns a four grid right maze. These are standardized coordinates.
}
\examples{
\dontrun{

# Returns a Grid with rank = 4
data(gridFourRight)
coordinates <- gridFourRight

}

}
\keyword{datasets}
