% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gridFourteenDown}
\alias{gridFourteenDown}
\title{Grid Fourteen Down}
\format{A data frame with 2 columns
\describe{
  \item{start}{Coordinates of Start Node.}
  \item{end}{Coordinates End Node.}
}}
\usage{
data(gridFourteenDown)
}
\description{
This returns a fourteen grid downards maze. These are standardized coordinates.
}
\examples{
\dontrun{

# Returns a Grid with rank = 14
data(gridFourteenDown)
coordinates <- gridFourteenDown

}

}
\keyword{datasets}
