% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gridFourteenLeft}
\alias{gridFourteenLeft}
\title{Grid Fourteen Left}
\format{A data frame with 2 columns
\describe{
  \item{start}{Coordinates of Start Node.}
  \item{end}{Coordinates End Node.}
}}
\usage{
data(gridFourteenLeft)
}
\description{
This returns a fourteen grid left maze. These are standardized coordinates.
}
\examples{
\dontrun{

# Returns a Grid with rank = 14
data(gridFourteenLeft)
coordinates <- gridFourteenLeft

}

}
\keyword{datasets}
