% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gridFourteenRight}
\alias{gridFourteenRight}
\title{Grid Fourteen Right}
\format{A data frame with 2 columns
\describe{
  \item{start}{Coordinates of Start Node.}
  \item{end}{Coordinates End Node.}
}}
\usage{
data(gridFourteenRight)
}
\description{
This returns a fourteen grid right maze. These are standardized coordinates.
}
\examples{
\dontrun{

# Returns a Grid with rank = 14
data(gridFourteenRight)
coordinates <- gridFourteenRight

}

}
\keyword{datasets}
