% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gridFourteenUp}
\alias{gridFourteenUp}
\title{Grid Fourteen Up}
\format{A data frame with 2 columns
\describe{
  \item{start}{Coordinates of Start Node.}
  \item{end}{Coordinates End Node.}
}}
\usage{
data(gridFourteenUp)
}
\description{
This returns a fourteen grid upwards maze. These are standardized coordinates.
}
\examples{
\dontrun{

# Returns a Grid with rank = 14
data(gridFourteenUp)
coordinates <- gridFourteenUp

}

}
\keyword{datasets}
