% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gridFourUp}
\alias{gridFourUp}
\title{Grid Four Up}
\format{A data frame with 2 columns
\describe{
  \item{start}{Coordinates of Start Node.}
  \item{end}{Coordinates End Node.}
}}
\usage{
data(gridFourUp)
}
\description{
This returns a four grid upwards maze. These are standardized coordinates.
}
\examples{
\dontrun{

# Returns a Grid with rank = 4
data(gridFourUp)
coordinates <- gridFourUp

}

}
\keyword{datasets}
