% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gridNineDown}
\alias{gridNineDown}
\title{Grid Nine Down}
\format{A data frame with 2 columns
\describe{
  \item{start}{Coordinates of Start Node.}
  \item{end}{Coordinates End Node.}
}}
\usage{
data(gridNineDown)
}
\description{
This returns a nine grid downwards maze. These are standardized coordinates.
}
\examples{
\dontrun{

# Returns a Grid with rank = 9
data(gridNineDown)
coordinates <- gridNineDown

}

}
\keyword{datasets}
