% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gridNineLeft}
\alias{gridNineLeft}
\title{Grid Nine Left}
\format{A data frame with 2 columns
\describe{
  \item{start}{Coordinates of Start Node.}
  \item{end}{Coordinates End Node.}
}}
\usage{
data(gridNineLeft)
}
\description{
This returns a nine grid left maze. These are standardized coordinates.
}
\examples{
\dontrun{

# Returns a Grid with rank = 9
data(gridNineLeft)
coordinates <- gridNineLeft

}

}
\keyword{datasets}
