% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gridNineRight}
\alias{gridNineRight}
\title{Grid Nine Right}
\format{A data frame with 2 columns
\describe{
  \item{start}{Coordinates of Start Node.}
  \item{end}{Coordinates End Node.}
}}
\usage{
data(gridNineRight)
}
\description{
This returns a nine grid right maze. These are standardized coordinates.
}
\examples{
\dontrun{

# Returns a Grid with rank = 9
data(gridNineRight)
coordinates <- gridNineRight

}

}
\keyword{datasets}
