% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gridNineteenUp}
\alias{gridNineteenUp}
\alias{gridNineteenUp}
\alias{gridNineteenDown}
\alias{gridNineteenUp}
\alias{gridNineteenLeft}
\alias{gridNineteenUp}
\alias{gridNineteenRight}
\title{Grid Nineteen Up}
\format{A data frame with 2 columns
\describe{
  \item{start}{Coordinates of Start Node.}
  \item{end}{Coordinates End Node.}
}}
\usage{
data(gridNineteenUp)

data(gridNineteenDown)

data(gridNineteenLeft)

data(gridNineteenRight)
}
\description{
This returns a nineteen grid right maze. These are standardized coordinates.

This returns a nineteen grid right maze. These are standardized coordinates.

This returns a nineteen grid right maze. These are standardized coordinates.

This returns a nineteen grid right maze. These are standardized coordinates.
}
\examples{
\dontrun{

# Returns a Grid with rank = 19
data(gridNineteenUp)
coordinates <- gridNineteenUp

}

\dontrun{

# Returns a Grid with rank = 19
data(gridNineteenDown)
coordinates <- gridNineteenDown

}

\dontrun{

# Returns a Grid with rank = 19
data(gridNineteenLeft)
coordinates <- gridNineteenLeft

}

\dontrun{

# Returns a Grid with rank = 19
data(gridNineteenRight)
coordinates <- gridNineteenRight

}

}
\keyword{datasets}
