% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gridNineUp}
\alias{gridNineUp}
\title{Grid Nine Up}
\format{A data frame with 2 columns
\describe{
  \item{start}{Coordinates of Start Node.}
  \item{end}{Coordinates End Node.}
}}
\usage{
data(gridNineUp)
}
\description{
This returns a nine grid upwards maze. These are standardized coordinates.
}
\examples{
\dontrun{

# Returns a Grid with rank = 9
data(gridNineUp)
coordinates <- gridNineUp

}

}
\keyword{datasets}
