% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gridSevenDown}
\alias{gridSevenDown}
\title{Grid Seven Down}
\format{A data frame with 2 columns
\describe{
  \item{start}{Coordinates of Start Node.}
  \item{end}{Coordinates End Node.}
}}
\usage{
data(gridSevenDown)
}
\description{
This returns a seven grid downwards maze. These are standardized coordinates.
}
\examples{
\dontrun{

# Returns a Grid with rank = 7
data(gridSevenDown)
coordinates <- gridSevenDown

}

}
\keyword{datasets}
