% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gridSevenLeft}
\alias{gridSevenLeft}
\title{Grid Seven Left}
\format{A data frame with 2 columns
\describe{
  \item{start}{Coordinates of Start Node.}
  \item{end}{Coordinates End Node.}
}}
\usage{
data(gridSevenLeft)
}
\description{
This returns a seven grid left maze. These are standardized coordinates.
}
\examples{
\dontrun{

# Returns a Grid with rank = 7
data(gridSevenLeft)
coordinates <- gridSevenLeft

}

}
\keyword{datasets}
