% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gridSevenRight}
\alias{gridSevenRight}
\title{Grid Seven Right}
\format{A data frame with 2 columns
\describe{
  \item{start}{Coordinates of Start Node.}
  \item{end}{Coordinates End Node.}
}}
\usage{
data(gridSevenRight)
}
\description{
This returns a seven grid right maze. These are standardized coordinates.
}
\examples{
\dontrun{

# Returns a Grid with rank = 7
data(gridSevenRight)
coordinates <- gridSevenRight

}

}
\keyword{datasets}
