% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gridSeventeenDown}
\alias{gridSeventeenDown}
\title{Grid Seventeen Down}
\format{A data frame with 2 columns
\describe{
  \item{start}{Coordinates of Start Node.}
  \item{end}{Coordinates End Node.}
}}
\usage{
data(gridSeventeenDown)
}
\description{
This returns a seventeen grid downards maze. These are standardized coordinates.
}
\examples{
\dontrun{

# Returns a Grid with rank = 17
data(gridSeventeenDown)
coordinates <- gridSeventeenDown

}

}
\keyword{datasets}
