% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gridSeventeenLeft}
\alias{gridSeventeenLeft}
\title{Grid Seventeen Left}
\format{A data frame with 2 columns
\describe{
  \item{start}{Coordinates of Start Node.}
  \item{end}{Coordinates End Node.}
}}
\usage{
data(gridSeventeenLeft)
}
\description{
This returns a seventeen grid left maze. These are standardized coordinates.
}
\examples{
\dontrun{

# Returns a Grid with rank = 17
data(gridSeventeenLeft)
coordinates <- gridSeventeenLeft

}

}
\keyword{datasets}
