% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gridSeventeenRight}
\alias{gridSeventeenRight}
\title{Grid Seventeen Right}
\format{A data frame with 2 columns
\describe{
  \item{start}{Coordinates of Start Node.}
  \item{end}{Coordinates End Node.}
}}
\usage{
data(gridSeventeenRight)
}
\description{
This returns a seventeen grid right maze. These are standardized coordinates.
}
\examples{
\dontrun{

# Returns a Grid with rank = 17
data(gridSeventeenRight)
coordinates <- gridSeventeenRight

}

}
\keyword{datasets}
