% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gridSeventeenUp}
\alias{gridSeventeenUp}
\title{Grid Seventeen Up}
\format{A data frame with 2 columns
\describe{
  \item{start}{Coordinates of Start Node.}
  \item{end}{Coordinates End Node.}
}}
\usage{
data(gridSeventeenUp)
}
\description{
This returns a seventeen grid upwards maze. These are standardized coordinates.
}
\examples{
\dontrun{

# Returns a Grid with rank = 17
data(gridSeventeenUp)
coordinates <- gridSeventeenUp

}

}
\keyword{datasets}
