% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gridSevenUp}
\alias{gridSevenUp}
\title{Grid Seven Up}
\format{A data frame with 2 columns
\describe{
  \item{start}{Coordinates of Start Node.}
  \item{end}{Coordinates End Node.}
}}
\usage{
data(gridSevenUp)
}
\description{
This returns a seven grid upwards maze. These are standardized coordinates.
}
\examples{
\dontrun{

# Returns a Grid with rank = 7
data(gridSevenUp)
coordinates <- gridSevenUp

}

}
\keyword{datasets}
