% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gridSixDown}
\alias{gridSixDown}
\title{Grid Six Down}
\format{A data frame with 2 columns
\describe{
  \item{start}{Coordinates of Start Node.}
  \item{end}{Coordinates End Node.}
}}
\usage{
data(gridSixDown)
}
\description{
This returns a six grid downwards maze. These are standardized coordinates.
}
\examples{
\dontrun{

# Returns a Grid with rank = 6
data(gridSixDown)
coordinates <- gridSixDown

}

}
\keyword{datasets}
