% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gridSixLeft}
\alias{gridSixLeft}
\title{Grid Six Left}
\format{A data frame with 2 columns
\describe{
  \item{start}{Coordinates of Start Node.}
  \item{end}{Coordinates End Node.}
}}
\usage{
data(gridSixLeft)
}
\description{
This returns a six grid left maze. These are standardized coordinates.
}
\examples{
\dontrun{

# Returns a Grid with rank = 6
data(gridSixLeft)
coordinates <- gridSixLeft

}

}
\keyword{datasets}
