% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gridSixRight}
\alias{gridSixRight}
\title{Grid Six Right}
\format{A data frame with 2 columns
\describe{
  \item{start}{Coordinates of Start Node.}
  \item{end}{Coordinates End Node.}
}}
\usage{
data(gridSixRight)
}
\description{
This returns a six grid right maze. These are standardized coordinates.
}
\examples{
\dontrun{

# Returns a Grid with rank = 6
data(gridSixRight)
coordinates <- gridSixRight

}

}
\keyword{datasets}
