% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gridSixteenDown}
\alias{gridSixteenDown}
\title{Grid Sixteen Down}
\format{A data frame with 2 columns
\describe{
  \item{start}{Coordinates of Start Node.}
  \item{end}{Coordinates End Node.}
}}
\usage{
data(gridSixteenDown)
}
\description{
This returns a sixteen grid downards maze. These are standardized coordinates.
}
\examples{
\dontrun{

# Returns a Grid with rank = 16
data(gridSixteenDown)
coordinates <- gridSixteenDown

}

}
\keyword{datasets}
