% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gridSixteenLeft}
\alias{gridSixteenLeft}
\title{Grid Sixteen Left}
\format{A data frame with 2 columns
\describe{
  \item{start}{Coordinates of Start Node.}
  \item{end}{Coordinates End Node.}
}}
\usage{
data(gridSixteenLeft)
}
\description{
This returns a sixteen grid left maze. These are standardized coordinates.
}
\examples{
\dontrun{

# Returns a Grid with rank = 16
data(gridSixteenLeft)
coordinates <- gridSixteenLeft

}

}
\keyword{datasets}
