% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gridSixteenRight}
\alias{gridSixteenRight}
\title{Grid Sixteen Right}
\format{A data frame with 2 columns
\describe{
  \item{start}{Coordinates of Start Node.}
  \item{end}{Coordinates End Node.}
}}
\usage{
data(gridSixteenRight)
}
\description{
This returns a sixteen grid right maze. These are standardized coordinates.
}
\examples{
\dontrun{

# Returns a Grid with rank = 16
data(gridSixteenRight)
coordinates <- gridSixteenRight

}

}
\keyword{datasets}
