% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gridSixteenUp}
\alias{gridSixteenUp}
\title{Grid Sixteen Up}
\format{A data frame with 2 columns
\describe{
  \item{start}{Coordinates of Start Node.}
  \item{end}{Coordinates End Node.}
}}
\usage{
data(gridSixteenUp)
}
\description{
This returns a sixteen grid upwards maze. These are standardized coordinates.
}
\examples{
\dontrun{

# Returns a Grid with rank = 16
data(gridSixteenUp)
coordinates <- gridSixteenUp

}

}
\keyword{datasets}
