% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gridTenDown}
\alias{gridTenDown}
\title{Grid Ten Down}
\format{A data frame with 2 columns
\describe{
  \item{start}{Coordinates of Start Node.}
  \item{end}{Coordinates End Node.}
}}
\usage{
data(gridTenDown)
}
\description{
This returns a ten grid downards maze. These are standardized coordinates.
}
\examples{
\dontrun{

# Returns a Grid with rank = 10
data(gridTenDown)
coordinates <- gridTenDown

}

}
\keyword{datasets}
