% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gridTenLeft}
\alias{gridTenLeft}
\title{Grid Ten Left}
\format{A data frame with 2 columns
\describe{
  \item{start}{Coordinates of Start Node.}
  \item{end}{Coordinates End Node.}
}}
\usage{
data(gridTenLeft)
}
\description{
This returns a ten grid left maze. These are standardized coordinates.
}
\examples{
\dontrun{

# Returns a Grid with rank = 10
data(gridTenLeft)
coordinates <- gridTenLeft

}

}
\keyword{datasets}
