% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gridTenRight}
\alias{gridTenRight}
\title{Grid Ten Right}
\format{A data frame with 2 columns
\describe{
  \item{start}{Coordinates of Start Node.}
  \item{end}{Coordinates End Node.}
}}
\usage{
data(gridTenRight)
}
\description{
This returns a ten grid right maze. These are standardized coordinates.
}
\examples{
\dontrun{

# Returns a Grid with rank = 10
data(gridTenRight)
coordinates <- gridTenRight

}

}
\keyword{datasets}
