% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gridTenUp}
\alias{gridTenUp}
\title{Grid Ten Up}
\format{A data frame with 2 columns
\describe{
  \item{start}{Coordinates of Start Node.}
  \item{end}{Coordinates End Node.}
}}
\usage{
data(gridTenUp)
}
\description{
This returns a ten grid upwards maze. These are standardized coordinates.
}
\examples{
\dontrun{

# Returns a Grid with rank = 10
data(gridTenUp)
coordinates <- gridTenUp

}

}
\keyword{datasets}
