% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gridThirteenDown}
\alias{gridThirteenDown}
\title{Grid Thirteen Down}
\format{A data frame with 2 columns
\describe{
  \item{start}{Coordinates of Start Node.}
  \item{end}{Coordinates End Node.}
}}
\usage{
data(gridThirteenDown)
}
\description{
This returns a thirteen grid downards maze. These are standardized coordinates.
}
\examples{
\dontrun{

# Returns a Grid with rank = 13
data(gridThirteenDown)
coordinates <- gridThirteenDown

}

}
\keyword{datasets}
