% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gridThirteenLeft}
\alias{gridThirteenLeft}
\title{Grid Thirteen Left}
\format{A data frame with 2 columns
\describe{
  \item{start}{Coordinates of Start Node.}
  \item{end}{Coordinates End Node.}
}}
\usage{
data(gridThirteenLeft)
}
\description{
This returns a thirteen grid left maze. These are standardized coordinates.
}
\examples{
\dontrun{

# Returns a Grid with rank = 13
data(gridThirteenLeft)
coordinates <- gridThirteenLeft

}

}
\keyword{datasets}
