% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gridThirteenRight}
\alias{gridThirteenRight}
\title{Grid Thirteen Right}
\format{A data frame with 2 columns
\describe{
  \item{start}{Coordinates of Start Node.}
  \item{end}{Coordinates End Node.}
}}
\usage{
data(gridThirteenRight)
}
\description{
This returns a thirteen grid right maze. These are standardized coordinates.
}
\examples{
\dontrun{

# Returns a Grid with rank = 13
data(gridThirteenRight)
coordinates <- gridThirteenRight

}

}
\keyword{datasets}
