% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gridThirteenUp}
\alias{gridThirteenUp}
\title{Grid Thirteen Up}
\format{A data frame with 2 columns
\describe{
  \item{start}{Coordinates of Start Node.}
  \item{end}{Coordinates End Node.}
}}
\usage{
data(gridThirteenUp)
}
\description{
This returns a thirteen grid upwards maze. These are standardized coordinates.
}
\examples{
\dontrun{

# Returns a Grid with rank = 13
data(gridThirteenUp)
coordinates <- gridThirteenUp

}

}
\keyword{datasets}
