% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gridThreeDown}
\alias{gridThreeDown}
\title{Grid Three Down}
\format{A data frame with 2 columns
\describe{
  \item{start}{Coordinates of Start Node.}
  \item{end}{Coordinates End Node.}
}}
\usage{
data(gridThreeDown)
}
\description{
This returns a three grid downwards maze. These are standardized coordinates.
}
\examples{
\dontrun{

# Returns a Grid with rank = X
data(gridThreeDown)
coordinates <- gridThreeDown

}

}
\keyword{datasets}
