% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gridThreeLeft}
\alias{gridThreeLeft}
\title{Grid Three Left}
\format{A data frame with 2 columns
\describe{
  \item{start}{Coordinates of Start Node.}
  \item{end}{Coordinates End Node.}
}}
\usage{
data(gridThreeLeft)
}
\description{
This returns a three grid left maze. These are standardized coordinates.
}
\examples{
\dontrun{

# Returns a Grid with rank = 3
data(gridThreeLeft)
coordinates <- gridThreeLeft

}

}
\keyword{datasets}
