% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gridThreeRight}
\alias{gridThreeRight}
\title{Grid Three Right}
\format{A data frame with 2 columns
\describe{
  \item{start}{Coordinates of Start Node.}
  \item{end}{Coordinates End Node.}
}}
\usage{
data(gridThreeRight)
}
\description{
This returns a three grid left maze. These are standardized coordinates.
}
\examples{
\dontrun{

# Returns a Grid with rank = 3
data(gridThreeRight)
coordinates <- gridThreeRight

}

}
\keyword{datasets}
