% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gridThreeUp}
\alias{gridThreeUp}
\title{Grid Three Up}
\format{A data frame with 2 columns
\describe{
  \item{start}{start, coordinates of Start Node.}
  \item{carat}{end, coordinates End Node.}
}}
\usage{
data(gridThreeUp)
}
\description{
This returns a three grid upwards maze. These are standardized coordinates.
}
\examples{
\dontrun{

# Returns a Grid with rank = 3
data(gridThreeUp)
coordinates <- gridThreeUp

}

}
\keyword{datasets}
