% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gridTwelveDown}
\alias{gridTwelveDown}
\title{Grid Twelve Down}
\format{A data frame with 2 columns
\describe{
  \item{start}{Coordinates of Start Node.}
  \item{end}{Coordinates End Node.}
}}
\usage{
data(gridTwelveDown)
}
\description{
This returns a twelve grid downards maze. These are standardized coordinates.
}
\examples{
\dontrun{

# Returns a Grid with rank = 12
data(gridTwelveDown)
coordinates <- gridTwelveDown

}

}
\keyword{datasets}
