% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gridTwelveLeft}
\alias{gridTwelveLeft}
\title{Grid Twelve Left}
\format{A data frame with 2 columns
\describe{
  \item{start}{Coordinates of Start Node.}
  \item{end}{Coordinates End Node.}
}}
\usage{
data(gridTwelveLeft)
}
\description{
This returns a twelve grid left maze. These are standardized coordinates.
}
\examples{
\dontrun{

# Returns a Grid with rank = 12
data(gridTwelveLeft)
coordinates <- gridTwelveLeft

}

}
\keyword{datasets}
