% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gridTwelveRight}
\alias{gridTwelveRight}
\title{Grid Twelve Right}
\format{A data frame with 2 columns
\describe{
  \item{start}{Coordinates of Start Node.}
  \item{end}{Coordinates End Node.}
}}
\usage{
data(gridTwelveRight)
}
\description{
This returns a twelve grid right maze. These are standardized coordinates.
}
\examples{
\dontrun{

# Returns a Grid with rank = 12
data(gridTwelveRight)
coordinates <- gridTwelveRight

}

}
\keyword{datasets}
