% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gridTwelveUp}
\alias{gridTwelveUp}
\title{Grid Twelve Up}
\format{A data frame with 2 columns
\describe{
  \item{start}{Coordinates of Start Node.}
  \item{end}{Coordinates End Node.}
}}
\usage{
data(gridTwelveUp)
}
\description{
This returns a twelve grid upwards maze. These are standardized coordinates.
}
\examples{
\dontrun{

# Returns a Grid with rank = 12
data(gridTwelveUp)
coordinates <- gridTwelveUp

}

}
\keyword{datasets}
