% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gridTwentyUp}
\alias{gridTwentyUp}
\title{Grid Twenty Up}
\format{A data frame with 2 columns
\describe{
  \item{start}{Coordinates of Start Node.}
  \item{end}{Coordinates End Node.}
}}
\usage{
data(gridTwentyUp)
}
\description{
This returns a twenty grid right maze. These are standardized coordinates.
}
\examples{
\dontrun{

# Returns a Grid with rank = 20
data(gridTwentyUp)
coordinates <- gridTwentyUp

}

}
\keyword{datasets}
