% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/howmany.R
\name{howMany}
\alias{howMany}
\title{howMany}
\usage{
howMany(rank, satPercent)
}
\arguments{
\item{rank}{This is the rank of the maze.}

\item{satPercent}{The percentage of saturation. Between 0-1.}
}
\description{
Calculate how many possible variation of black dotes for a given saturation.
}
\details{
Calculate how many possible variation of black dotes for a given saturation. The first node will not be a black dot.
}
\examples{
howMany(rank=5, satPercent=0.5)
}
\seealso{
\code{\link{lowerGrid}}
}
\author{
Aiden Loe
}
\keyword{lowerGrid}
