% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowerGrid.R
\name{lowerGrid}
\alias{lowerGrid}
\title{lowerGrid}
\usage{
lowerGrid(rank = 5)
}
\arguments{
\item{rank}{This is the rank of the maze.}
}
\description{
This tells you all the node position in the maze.
}
\details{
The construction of the maze is first created in a symmetrical format. However, only half of the nodes are kept in order to create the actual maze. Hence, this function calculates the nodePosition of the actual maze.
}
\examples{
lowerGrid(3)
}
\author{
Aiden Loe
}
