% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colourNodePosition.R
\name{np}
\alias{np}
\title{Colour Node Position}
\usage{
np(rank = 3, satPercent = 0.5, seed = 1)
}
\arguments{
\item{rank}{This is the rank of the maze.}

\item{satPercent}{Percentage of saturation.}

\item{seed}{To always get the same position for a local computer.}
}
\value{
A 'np' class which will be used for other functions in the package.
}
\description{
Returns the colour node position. You need to use the node position function first.
}
\details{
This function will not sample from the first node position. If you consider sampling from the first node, then in javascript, the summing of the black dotes need to begin from 1 rather than 0. To keep it simple, always ensure that the first node is not sampled as a black dot.
}
\examples{

np(rank=3,satPercent=0.5,seed=1)

}
\seealso{
\code{\link{mazeEst}}, \code{\link{genPathSeed}}
}
\author{
Aiden Loe
}
