% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topNodes.R
\name{topNodes}
\alias{topNodes}
\title{Top Nodes}
\usage{
topNodes(rank)
}
\arguments{
\item{rank}{This is the Rank of the maze.}
}
\description{
The node length calculates all the nodes on the longest row for a given rank.
}
\details{
This needs to have a rank value of greater than 1. This is needed so that you can cross check how many coloured nodes are located on the longest row.
}
\examples{
rank <-3
topNodes(rank)
}
\author{
Aiden Loe
}
